backend/requirements/requirements.txt
Django>=4.1,<5.0
django
djangorestframework
psycopg2-binary
django-cors-headers
boto3
django-storages
